Use epoch + sequence number
DTLS uses epoch concatenated with the record sequence number for
in place of the intrinsic 64-bit sequence number in TLS. Use this
in encrypt(), decrypt(), and within the handshake.

Still getting an 'unexpected message' alert from browser. Wrote
a tool, generate_cert.go, go generating a certificate from a
template, client.crt, which is from a real WebRTC session. May
work better than OpenSSL generated certificate, which is signed
with SHA-1 but WebRTC is signed with SHA-256.