Retry newSessionTicketMsg reads in DTLS handshake
In cases where ServerHello is retransmitted as ClientCertificate
was not received in time, the second ServerHello is processed
when a NewSessionTicket is expected. Retry reading NewSessionTicket
in these cases.

This is a simple hack, stemming from porting TLS to DTLS. The proper
fix is to replace stream-based reads/writes used in TCP based TLS
with DTLS's state machine, as described in RFC 4347.