Discard incoming packets if they are not being read in time.
Both ICE messages and RTP/RTCP are multiplexed over a single UDP socket.
The ICE agent must therefore dequeue and process every packet that comes
over the socket. If the RTP/RTCP packet are not being processed quickly
enough, they must be discarded so that ICE can continue to operate.