Add -6 flag to toggle use of IPv6 (default disabled).
ICE prefers IPv6 addresses, all else being equal. I've noticed however
when testing locally that when IPv6 addresses are used, sometimes the
connection is established but no data flows. Very likely this is just a
bug in logic for dealing with multiple bases, but for now the easiest
solution is to disable IPv6 by default.