Basic groundwork for generalizing media sources.
The goal is to start preparing for audio and/or other video formats,
while also solving the problem of V4L2 not compiling on darwin.

The set of allowed source types is determined at runtime, with each
supported source being registered in an `init()`. To stream video using
v4l2 on the RPi, specify `-i v4l2:/dev/video0` on the command line (this
is the default). To stream from a raw H.264 file (or pipe), use
`-i h264:path/to/file.264`.