Handle disconnections (#50)
* Handle disconnections

Use a teardown signalling channel to stop video when the peer
connection is closed. Close peer connection when websocket
connection is closed by moving defer out of goroutine and into
websocket handler.

* Do not panic if cannot open video device

If video is in use (another window is open), do not need to panic
and exit -- simply return to the websocket handler.