Split signaling logic out of alohacam package.
The goal here is to wrap up the signaling flow into a simple interface,
so that
 1. it is more clear exactly where and how we rely on signaling, and
 2. we can add other signaling implementations without introducing a
separate `main` package for each one.

The top-level interface here is a `signaling.Client`, which encapsulates
all the signaling logic. The `Listen()` function just waits for a remote
peer to initiate a new session. The signaling client then invokes the
`signaling.SessionHandler`, passing it the necessary channels and hooks
to perform offer/answer negotiation and ICE candidate exchange.

For now there's only a single signaling implementation, which is the one
from the `alohacam` demo that runs a local webserver. This is actually a
degenerate case, where the signaling client also performs the role of a
signaling *server*.

The static files are bundled into the binary using the `esc` tool, which
is relatively lightweight but serves our use case very well. Install
with `go get github.com/mjibson/esc`, then run `make generate`.