Clean up agent goroutines on teardown
Propagate context to demuxStun, which runs on a separate goroutine,
to be able to terminate the goroutine if blocked on writing a
non-STUN packet to channel.

Close base upon exit from agent loop upon teardown. Closing the
base closes the underlying UDPConn, which causes pending ReadFrom
calls to return with an error. That error is _not_ io.EOF, which
is generated when the remote closes the connection, not possible
on a stateless connection such as UDPConn. Instead, a net.OpError
is returned. Handle this error.