Implement triggered connectivity checks.
Triggered checks are intended to circumvent the normal ordering of
connectivity checks, such that if our ICE agent receives a connectivity
check from the remote peer, it immediately tries to send its own along
the same network path. This isn't necessary for ICE to function
properly, but it's a nice optimization. See
https://tools.ietf.org/html/rfc8445#section-6.1.4

This fixes issue #73 (peer reflexive candidates getting pruned) in a
roundabout way. Since the peer reflexive candidate pair gets added to
the triggered check queue, it still exists in memory even though it gets
pruned from the checklist. So if the triggered check succeeds, the pair
gets added to the valid list, and can therefore be selected even though
it was pruned. Not very elegant, but it works.

Also remove default case when writing to the checklist listener
channels. Otherwise I noticed that sometimes the checklist update would
be missed.