Implement per-Endpoint packet buffering for Mux.
Each `Endpoint` maintains a circular queue of packet buffers (a buffer
of buffers). When the `Mux` read loop receives a packet, it trades the
packet buffer for one of the target `Endpoint`'s buffers. This "give a
penny, take a penny" approach allows each `Endpoint` to have its own
queue of buffered packets, without having to copy each packet between
the `Mux` and the `Endpoint`.

If an `Endpoint` has no readers, or if the readers are not fast enough,
its packets will be queued up to certain count (currently 32). Beyond
that, new incoming packets will replace the oldest packets, which gives
a slow reader the best possible chance to catch up.