Allow Buffer to be closed.
This is necessary to be able to unblock `Buffer` readers that are
waiting for the next packet.

Note: The way `Buffer` is currently used it acts much like `io.Pipe`,
except that it allows packets to be buffered. There's probably a way to
achieve this without requiring an allocation on every `Write`. See also
https://github.com/golang/go/issues/28790.