Support multiple network interfaces.
The local ICE agent nows runs a single checklist loop, rather than a
separate checklist loop per base. This involved a lot of internal
refactoring of the `ice` package.

The public interface for `ice.Agent` now makes goroutine accounting more
clear. Every function that spawns long-running goroutines takes a
`context.Context` argument, and those goroutines should terminate when
the context is canceled. The new `ice.DataStream` type represents the
result of ICE processing, a bidirectional data stream to the remote
peer.

Also simplify the public interface for `PeerConnection`, to emphasize
the three major stages in establishing a peer connection:
 1. SDP offer/answer exchange (`SetRemoteDescription`)
 2. ICE candidate exchange (`ExchangeCandidates`)
 3. Connection and media streaming (`Connect`)
The call to `Connect()` is blocking. It terminates if a connection
cannot be established, or if the underlying UDP connection times out, or
when `Close()` is called explicitly.