Update CircleCI configuration for Go modules.
With `GO111MODULE=auto`, module support is disabled for packages within
`GOPATH`. So we just check out the repo to the default working directory
(`~/project`).

Also, we don't need to explicitly `go get` dependencies, since that will
happen automatically when we `go build`. (In the future, if we want to
cache dependencies, we could use `go mod download`.)

However, the `esc` and `richgo` tools still need to be installed
explicitly. These `go get` commands will have the undesired side effect
of munging the local `go.mod` file (because Go modules are a mess), but
that shouldn't have any real consequences.