Make unsupport STUN methods non-fatal
Unsupported STUN methods, such as the Send method (method 6)
introduced by TURN and sent by TURN relays, should not be
fatal.

In particular, TURN relays send a STUN binding indication via
the Send method to enable clients to authenticate the relay.
The indication does not require a reply, and does not even need
to be handled to be able to send video via the relay, but it
must not terminate the program.