Tweak loopvideo.sh script for smoother playback.
The `keyint=25` option makes the encoder send a key frame (IDR picture)
every 25 frames, or roughly once per second. (The default for libx264 is
to send key frames only on "scene cuts".) This seems to be important for
browsers to be able to render the video.

The other options limit the frame rate to 2 Mbps, which helps prevent
stuttering (at least on my machine).