Support aggressive nomination of ICE candidates.
RFC 5245 (the original ICE spec) allows for a mode called "aggressive
nomination" (§8.1.1.2) , where the controlling agent nominates multiple
candidate pairs with the understanding that the highest-priority pair
for which connectivity checks succeed will be the one used. The
alternative is "regular nomination", where the controlling agent waits
for one or more pairs to pass connectivity checks before nominating a
single one.

RFC 8445 deprecates aggressive nomination (§8.1.1), and encourages
agents to to set the `ice-options:ice2` attribute to indicate that it
should not be used by the remote agent. However, it still requires that
agents be prepared to receive data on *any* valid pair. The changes here
should allow that, as well as supporting aggressive nomination.