Subscriber interface; Broadcaster
Adds a Subscriber interface plus Broadcaster component which
implements it. The MediaSourcer interface also inherts
Subscriber. The idea is Broadcaster is generally useful for
broadcasting media frames to multiple subscribers (multiple
viewers) with zero-copy -- media sources must implement this
interface, which enables PeerConnection goroutines to subscribe
to media sources provided as a Config object.

The idea is that sources also automatically encode the media
based on what Encoder is set on the source, although this is
not yet complete. It has the advantage of encoding once per
source, rather than once per subscriber, so lower compute,
but if two subscribers need different codecs, this won't
work. Fortunately, currently it does not look like they will.