Partial support for mdns-ice-candidates proposal.
https://tools.ietf.org/html/draft-ietf-rtcweb-mdns-ice-candidates-04

This handles the case where the browser generates an ephemeral mDNS
hostname (of the form "<uuid4>.local") and uses it as a host candidate
in place of an actual IP address. As of today the major browsers (namely
Chrome and Firefox) are doing this, as a privacy measure to avoid
leaking the browser's LAN IP address.

For our purposes it doesn't *really* matter, since `alohartcd` is still
sending *its* local IP address as a host candidate (which the browser
uses to immediately establish a connection). If we want to be extra
sensitive to privacy concerns, we could generate an ephemeral mDNS
hostname of our own and send that as our host candidate. This commit
provides a (currently unused) `mdnsAnnounce` function for that purpose.

(An alternative would be to avoid sending a host candidate altogether,
and instead rely on the browser to send theirs. The device's IP would
then appear only as a peer-reflexive candidate to the browser.)

All the mDNS logic currently lives in the `mdns.go` file. It might make
more sense for it to live in its own `internal/ice/mdns` package.